pyproject.toml
[project]
name = "identity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.48",
]

[dependency-groups]
dev = [
    "cryptography>=46.0.5",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pyjwt>=2.11.0",
    "uvicorn[standard]>=0.41.0",
]
